perbaikan fitur delete pada modal

+++ app/Http/Controllers/Prodi/KurikulumController.php
@@ -13,7 +13,7 @@ class KurikulumController extends Controller
 {
     function __construct()
     {
-        // $this->middleware('permission:read kurikulums', ['only' => ['index','show']]);
+        $this->middleware('permission:read kurikulums', ['only' => ['index','show']]);
         $this->middleware('permission:create kurikulums', ['only' => ['create','store']]);
         $this->middleware('permission:update kurikulums', ['only' => ['edit','update']]);
         $this->middleware('permission:delete kurikulums', ['only' => ['destroy']]);
@@ -26,8 +26,7 @@ class KurikulumController extends Controller
 
     public function create(Prodi $prodi)
     {
-        $kurikulum = new Kurikulum();
-        return view('setting.kurikulum-form', $this->_dataSelection($prodi, $kurikulum));
+        return to_route('home')->with('warning', 'Gunakan tombol tambah (modal) pada halaman Dashboard Prodi.');
     }
 
     public function store(Request $request)
@@ -40,7 +39,7 @@ class KurikulumController extends Controller
 
     public function edit(Prodi $prodi, Kurikulum $kurikulum)
     {
-        return view('setting.kurikulum-form', $this->_dataSelection($prodi, $kurikulum));
+        return to_route('home')->with('warning', 'Gunakan tombol edit (modal) pada daftar Kurikulum.');
     }
 
     public function update(Request $request, Prodi $prodi, Kurikulum $kurikulum)
@@ -55,6 +54,20 @@ class KurikulumController extends Controller
     public function destroy(Prodi $prodi, Kurikulum $kurikulum)
     {
         $name = strtoupper($kurikulum->name);
+
+        $isUsed = $kurikulum->profils()->exists()
+            || $kurikulum->cpls()->exists()
+            || $kurikulum->bks()->exists()
+            || $kurikulum->mks()->exists()
+            || $kurikulum->joinProfilCpls()->exists()
+            || $kurikulum->joinCplBks()->exists()
+            || $kurikulum->joinBkMks()->exists()
+            || $kurikulum->joinMkUsers()->exists();
+
+        if ($isUsed) {
+            return to_route('home')->with('error','Kurikulum '.$name.' tidak dapat dihapus karena sudah digunakan pada tabel relasi.');
+        }
+
         $kurikulum->delete();
         return to_route('home')->with('warning','Kurikulum '.$name.' telah dihapus');
     }
